--- admin/src/components/PluginIcon/index.js
@@ -1,12 +1,0 @@
-/**
- *
- * PluginIcon
- *
- */
-
-import React from "react";
-import { Download } from "@strapi/icons";
-
-const PluginIcon = () => <Download />;
-
-export default PluginIcon;
