+++ admin/src/pluginId.js
@@ -1,5 +1,1 @@
-import pluginPkg from '../../package.json';
-
-const pluginId = pluginPkg.name.replace(/^(@[^-,.][\w,-]+\/|strapi-)plugin-/i, '');
-
-export default pluginId;
+export const PLUGIN_ID = 'export-excel';
