--- admin/src/translations/fr.json
@@ -1,1 +1,0 @@
-{}
