--- admin/src/utils/getTrad.js
@@ -1,5 +1,0 @@
-import pluginId from '../pluginId';
-
-const getTrad = (id) => `${pluginId}.${id}`;
-
-export default getTrad;
