+++ admin/src/utils/getTranslation.js
@@ -1,0 +1,5 @@
+import { PLUGIN_ID } from '../pluginId';
+
+const getTranslation = (id) => `${PLUGIN_ID}.${id}`;
+
+export { getTranslation };
