+++ package.json
@@ -1,59 +1,64 @@
 {
-  "name": "strapi-plugin-excel-export-2024",
-  "version": "1.0.7",
-  "description": "A strapi plugin used which can be used to download data in xlsx format",
-  "strapi": {
-    "name": "excel-export",
-    "description": "A strapi plugin used which can be used to download data in xlsx format",
-    "kind": "plugin",
-    "displayName": "Excel Export"
+  "version": "0.0.0",
+  "keywords": [],
+  "type": "commonjs",
+  "exports": {
+    "./package.json": "./package.json",
+    "./strapi-admin": {
+      "source": "./admin/src/index.js",
+      "import": "./dist/admin/index.mjs",
+      "require": "./dist/admin/index.js",
+      "default": "./dist/admin/index.js"
+    },
+    "./strapi-server": {
+      "source": "./server/src/index.js",
+      "import": "./dist/server/index.mjs",
+      "require": "./dist/server/index.js",
+      "default": "./dist/server/index.js"
+    }
+  },
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "strapi-plugin build",
+    "watch": "strapi-plugin watch",
+    "watch:link": "strapi-plugin watch:link",
+    "verify": "strapi-plugin verify"
   },
   "dependencies": {
-    "@strapi/design-system": "^1.6.3",
-    "@strapi/helper-plugin": "^4.6.0",
-    "@strapi/icons": "^1.6.3",
-    "prop-types": "^15.7.2",
-    "react-data-table-component": "^7.6.2",
-    "exceljs": "^4.4.0"
+    "@strapi/design-system": "^2.0.0-rc.18",
+    "@strapi/icons": "^2.0.0-rc.18",
+    "axios": "^1.8.1",
+    "exceljs": "^4.4.0",
+    "react-data-table-component": "^7.7.0",
+    "react-intl": "^7.1.6"
   },
   "devDependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.4",
-    "styled-components": "^5.3.6"
+    "@strapi/sdk-plugin": "^5.3.2",
+    "@strapi/strapi": "^5.10.4",
+    "prettier": "^3.5.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.30.0",
+    "styled-components": "^6.1.15"
   },
   "peerDependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.4",
-    "styled-components": "^5.3.6",
-    "@strapi/strapi": "^4.0.0"
+    "@strapi/sdk-plugin": "^5.3.2",
+    "@strapi/strapi": "^5.10.4",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.30.0",
+    "styled-components": "^6.1.15"
   },
-  "author": {
-    "name": "Vinu Babu",
-    "email": "vinubabu323@gmail.com"
+  "strapi": {
+    "kind": "plugin",
+    "name": "export-excel",
+    "displayName": "",
+    "description": ""
   },
-  "engines": {
-    "node": ">=14.19.1 <=20.x.x",
-    "npm": ">=6.0.0"
-  },
+  "name": "export-excel",
+  "description": "",
   "license": "MIT",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/vinubabu323/strapi-plugin-excel-export.git"
-  },
-  "keywords": [
-    "strapi",
-    "strapi-excel-download",
-    "excel",
-    "excel-download",
-    "excel-export"
-  ],
-  "bugs": {
-    "url": "https://github.com/vinubabu323/strapi-plugin-excel-export/issues"
-  },
-  "homepage": "https://github.com/vinubabu323/strapi-plugin-excel-export#readme"
+  "author": "Extreeeme <martindcontact@gmail.com>"
 }
