--- server/config/index.js
@@ -1,6 +1,0 @@
-'use strict';
-
-module.exports = {
-  default: {},
-  validator() {},
-};
