--- server/controllers/index.js
@@ -1,7 +1,0 @@
-'use strict';
-
-const myController = require('./my-controller');
-
-module.exports = {
-  myController,
-};
