--- server/services/index.js
@@ -1,7 +1,0 @@
-'use strict';
-
-const myService = require('./my-service');
-
-module.exports = {
-  myService,
-};
