--- server/services/my-service.js
@@ -1,7 +1,0 @@
-'use strict';
-
-module.exports = ({ strapi }) => ({
-  getWelcomeMessage() {
-    return 'Welcome to Strapi 🚀';
-  },
-});
