+++ server/src/bootstrap.js
@@ -1,0 +1,5 @@
+const bootstrap = ({ strapi }) => {
+  // bootstrap phase
+};
+
+export default bootstrap;
