+++ server/src/config/index.js
@@ -1,0 +1,4 @@
+export default {
+  default: {},
+  validator() {},
+};
