+++ server/src/content-types/index.js
@@ -1,0 +1,1 @@
+export default {};
