+++ server/src/controllers/index.js
@@ -1,0 +1,5 @@
+import controller from './controller';
+
+export default {
+  controller,
+};
