+++ server/src/destroy.js
@@ -1,0 +1,5 @@
+const destroy = ({ strapi }) => {
+  // destroy phase
+};
+
+export default destroy;
