+++ server/src/register.js
@@ -1,0 +1,5 @@
+const register = ({ strapi }) => {
+  // register phase
+};
+
+export default register;
