+++ server/src/routes/index.js
@@ -1,0 +1,10 @@
+import contentAPIRoutes from './content-api';
+
+const routes = {
+  'content-api': {
+    type: 'content-api',
+    routes: contentAPIRoutes,
+  },
+};
+
+export default routes;
