+++ server/src/services/index.js
@@ -1,0 +1,5 @@
+import service from './service';
+
+export default {
+  service,
+};
