+++ server/src/services/service.js
@@ -1,0 +1,7 @@
+const service = ({ strapi }) => ({
+  getWelcomeMessage() {
+    return 'Welcome to Strapi 🚀';
+  },
+});
+
+export default service;
